move permission and fix scope
